ad9467: rename files using lower-case characters
Change made to abide to the Linux naming standards. Changed the include
file to keep the functionality. No actual functional changes.
